指摘事項の修正

+++ apps/back/src/main/java/io/github/taichi0373/benefit_map/repository/entity/FareDiscountEligibilityEntity.java
@@ -45,7 +45,7 @@ public class FareDiscountEligibilityEntity implements Serializable {
 
     /** 割引値 */
     @Column(name = "discount_value")
-    private BigDecimal discountValue;
+    private Integer discountValue;
 
     // -------------------------------------------------------------------------
     // 利用条件情報（絞り込みに使用）
